fixed error with mutation of state during showing the progress-bar for uploading file

+++ front_end/src/components/disk/fileList/file/File.jsx
@@ -29,7 +29,9 @@ const File = ({file}) => {
 
 	function deleteClickHandler(e){
 		e.stopPropagation()
-		dispatch(deleteFile(file))
+		if (window.confirm("Are you sure you want to delete file?") === true) {
+			dispatch(deleteFile(file))
+		}
 	}
 
 	if (fileView === 'list') {
